# core/openapi/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StarterKit API Documentation",
    "description": "API documentation using OpenAPI specification for Next.js Starter Kit",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:6001",
      "description": "Local Next.js Dev Server"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with email and password",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token using refresh token",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TokenRefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenRefreshResponse" }
              }
            }
          },
          "401": {
            "description": "Refresh token expired or invalid"
          }
        }
      }
    },
    "/api/home/guide-cards": {
      "get": {
        "tags": ["Home"],
        "summary": "Get list of guide cards",
        "description": "Returns a list of guide cards for the home page, each containing an ID, thumbnail image, and video source.",
        "operationId": "getGuideCards",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideCardCollection"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "description": "Authenticated user",
        "required": ["id", "email", "name", "role"],
        "properties": {
          "id": { "type": "string", "example": "usr_123" },
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "name": { "type": "string", "example": "John Doe" },
          "role": { "type": "string", "enum": ["admin", "user"], "example": "user" }
        },
        "type": "object"
      },
      "LoginRequest": {
        "description": "Login credentials",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "example": "password123" }
        },
        "type": "object"
      },
      "AuthResponse": {
        "description": "Successful auth response with tokens and user",
        "required": ["user", "accessToken", "refreshToken"],
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" }
        },
        "type": "object"
      },
      "TokenRefreshRequest": {
        "description": "Token refresh request",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string" }
        },
        "type": "object"
      },
      "TokenRefreshResponse": {
        "description": "New access token after refresh",
        "required": ["accessToken"],
        "properties": {
          "accessToken": { "type": "string" }
        },
        "type": "object"
      },
      "GuideCardResource": {
        "description": "A single guide card item",
        "required": ["id", "thumbnailSrc", "videoSrc"],
        "properties": {
          "id": {
            "type": "string",
            "example": "guide-1"
          },
          "thumbnailSrc": {
            "type": "string",
            "format": "uri",
            "example": "https://picsum.photos/seed/guide1/600/340"
          },
          "videoSrc": {
            "type": "string",
            "format": "uri",
            "example": "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4"
          }
        },
        "type": "object"
      },
      "GuideCardCollection": {
        "description": "List of guide cards",
        "properties": {
          "data": {
            "description": "Array of guide card items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuideCardResource"
            }
          }
        },
        "type": "object"
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Authentication endpoints"
    },
    {
      "name": "Home",
      "description": "Home page data endpoints"
    }
  ]
}